bot/pyproject.toml
[project]
name = "quiz_service_bot"
version = "0.1.0"
description = ""
authors = [
    {name = "KamranYsupov",email = "kamrankek48@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"

[tool.poetry]
packages = [{include = "q", from = "src"}]


[tool.poetry.dependencies]
aiogram = "^3.20.0.post0"
dependency-injector = "^4.46.0"
watchdog = "^6.0.0"
aiohttp = "^3.11.18"
loguru = "0.6"
python-dotenv = "^1.1.0"
celery = "^5.5.2"
redis = "^6.0.0"
opentelemetry-api = "^1.33.0"
opentelemetry-sdk = "^1.33.0"
opentelemetry-exporter-jaeger = "^1.21.0"
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
